pyproject.toml
[project]
name = "coderclient"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Peder Landsverk", email = "peder.landsverk@soprasteria.com" }
]
dependencies = [
    "requests>=2.32.2",
    "click>=8.1.7",
    "pydantic>=2.7.1",
    "pydantic-settings>=2.2.1",
    "jinja2>=3.1.4",
    "pyyaml>=6.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
coderclient="coderclient.entrypoint:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/coderclient"]
